# Refactor `Optimizer` to use owned plans and `TreeNode` API (10% faster planning)