@alamb 2024-04-09
@jackwener since you implemented some of the original optimizer recursion I wonder if you would have some time to review this PR